Change  the attitude to do everithing with 777 in linux
The creator of this tutorial must be a Windwos user. Just working with admin ;)

A tutoral with the title permission and ownership should at least not learn such things ... it is not a windows where we put admin as soon we get an error. As mentioned on the onion we work with root ... 755 is more than good.